Financial Babaji Web/ml_backend/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
pydantic==2.7.1
numpy==1.26.4
joblib==1.4.2
scikit-learn==1.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
